+++ src/symbols.cpp
@@ -477,7 +477,7 @@ uint64_t symbolResolverGetAddressID(uintptr_t _resolver, uint64_t _address, bool
 		{
 			Module& module = info->m_modules[i];
 			loadPDB(module, info->m_symbolStore);
-			uintptr_t id = module.m_PDBFile->getSymbolID(_address - module.m_module.m_baseAddress);
+			uint64_t id = module.m_PDBFile->getSymbolID(_address - module.m_module.m_baseAddress);
 			if (_isRTMdll)
 				*_isRTMdll = module.m_isRTMdll;
 			return id + module.m_module.m_baseAddress;
